Create and connect redux store
