+++ frontend/src/actions/types.js
@@ -1,0 +1,3 @@
+export const CORRECT_ANSWER = "CORRECT_ANSWER"
+export const INCORRECT_ANSWER = "INCORRECT_ANSWER"
+export const RESET_GAME = "RESET_GAME"
