+++ frontend/src/config.js
@@ -1,0 +1,2 @@
+export const TIMEOUT_BEFORE_NEXT_QUESTION = 600;
+export const MAX_LIVES = 3;
